backend/requirements.txt
Flask
Flask-CORS
tensorflow
numpy
opencv-python
pymongo
gunicorn
matplotlib
dotenv
pillow
werkzeug
